+++ examples/Controller/User.php
@@ -8,6 +8,7 @@ use WoohooLabs\Yin\Examples\JsonApi\Resource\ContactResourceTransformer;
 use WoohooLabs\Yin\Examples\JsonApi\Resource\UserResourceTransformer;
 use WoohooLabs\Yin\Examples\Repository\UserRepository;
 use WoohooLabs\Yin\JsonApi\Request\Request;
+use WoohooLabs\Yin\JsonApi\Response\FetchResponse;
 
 class User
 {
@@ -28,6 +29,6 @@ class User
 
         $document = new UserDocument(new UserResourceTransformer(new ContactResourceTransformer()));
 
-        return $document->getResponse($response, $resource, Request::fromServerRequest($request));
+        return FetchResponse::ok(Request::fromServerRequest($request), $response, $document, $resource);
     }
 }